gRest updates for db-sync v13 (#1437)
SQL Updates to start supporting dbsync v13, note that dbsync.sh / config changes were tackled via #1433.

  - columns added `inlineDatumId` and `referenceScriptId` (foreign keys for datumId and scriptId tables)
  - consider what to do with those

  - consider what to do with new tables

  - `reward_addr` was used raw in `pool_info_cache.sql` line 325, to be changed to use new reference `reward_addr_id` (address lookup required)

  - `pool_info_cache.sql` line 104 uses removed column `registered_tx_id`
  - `pool_info_cache.sql` line 104 uses removed columns `registered_tx_id` and `pool_hash_id`
  - needs to get that data for cache using new `pool_update_id` reference (lookup required)

  - `script_redeemers.sql` line 35 uses removed column `datum_id`
  - `tx_info.sql` line 501 uses removed column `datum_id`
  - this is replaced by reference to new table (RedeemerData)

  - consider additional uses for the new table, script endpoints are fairly limited

  - consider use for new table and if inclusion is necessary to `tx_info`

  - renamed column `coins_per_utxo_word` used in `tx_info.sql`

  - renamed column `coins_per_utxo_word` used in `epoch_info_cache.sql`
  - renamed column `entropy` used in `epoch_info_cache.sql` (this column was NOT renamed in `param_proposal` table)

  - consider use of new column `hash`

  - Include `proto_major` and `proto_minor` in /blocks view and /block_info endpoint

Closes cardano-community/koios-artifacts#53

Schema PR: cardano-community/koios-artifacts#55

+++ files/grest/rpc/assets/asset_info.sql
@@ -8,7 +8,7 @@ CREATE FUNCTION grest.asset_info (_asset_policy text, _asset_name text default '
     total_supply text,
     mint_cnt bigint,
     burn_cnt bigint,
-    creation_time double precision,
+    creation_time numeric,
     minting_tx_metadata json,
     token_registry_metadata json
   )
